refactor: move non-public code to internal folder
